it works!

+++ Mcu/h563/Inc/spi.h
@@ -36,10 +36,10 @@ typedef struct {
     // circular buffers
     // buffer size MUST be 256
     // implementation takes advantage of integer overflow
-    uint8_t* _rx_buffer;
-    uint8_t* _tx_buffer;
-    uint16_t _rx_buffer_size;
-    uint16_t _tx_buffer_size;
+    uint16_t* _rx_buffer;
+    uint16_t* _tx_buffer;
+    uint8_t _rx_buffer_size;
+    uint8_t _tx_buffer_size;
     uint8_t _rx_head;
     uint8_t _tx_head;
     uint8_t _tx_tail;
